Construct HTML elements and style header in CSS
A first edition of the landing page with focus on getting every element of html in the index.html file and styling the header. The header image is a statically sized picture that doesn't change dimensions when the window is scaled. This might be subject for change later to ensure that the height of the image matches the corresponding div to the left, but this requires cropping to avoid distortion, which I have yet to learn.